[FEATURE] added exclude classes [FEATURE] added some more config options to importer [BUGFIX] fixed return types for collections (array vs. ObjectStorage)

+++ Classes/Command/ImportCommand.php
@@ -7,6 +7,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Filesystem\Exception\FileNotFoundException;
+use TYPO3\CMS\Core\Core\Environment;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -24,6 +25,10 @@ use Ujamii\OpenImmoTypo3\Domain\Model\Openimmo;
  */
 class ImportCommand  extends Command
 {
+
+    const ASSET_TABLE_NAME = 'tx_openimmotypo3_domain_model_daten';
+    // TODO: this has to be "pfad" later...
+    const ASSET_FIELD_NAME = 'anhanginhalt';
 
     /**
      * @var InputInterface
@@ -47,8 +52,11 @@ class ImportCommand  extends Command
     {
         $this->setDescription('Imports openimmo xml data into the local database.');
         $this->addArgument('pid', InputArgument::REQUIRED, 'Target pid where the data shall be stored.');
-        $this->addArgument('sourceFolder', InputArgument::OPTIONAL, 'Source folder where the most recent *.zip file will be searched in.', '/uploads/tx_openimmo/');
+        $this->addArgument('sourceFolder', InputArgument::OPTIONAL, 'Source folder where the *.zip file will be searched in.', 'public/uploads/tx_openimmo/');
+        $this->addArgument('targetFolder', InputArgument::OPTIONAL, 'Target folder where data will be extrcated to.', 'public/uploads/tx_openimmo/');
         $this->addArgument('truncateTables', InputArgument::OPTIONAL, 'Whether the openimmo db tables shall be truncated before the import.', true);
+        $this->addArgument('adjustAssetPaths', InputArgument::OPTIONAL, 'Whether the paths of the binary files shall be set.', true);
+        $this->addArgument('removeZipfile', InputArgument::OPTIONAL, 'Whether the zip file shall be deleted after the import.', true);
     }
 
     /**
@@ -57,6 +65,7 @@ class ImportCommand  extends Command
      *
      * @return int|void|null
      * @throws \ReflectionException
+     * @throws \Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
@@ -64,13 +73,13 @@ class ImportCommand  extends Command
         $this->output = $output;
 
         // find and extract zip file
-        $sourceFolder = \TYPO3\CMS\Core\Core\Environment::getPublicPath() . $this->input->getArgument('sourceFolder');
+        $sourceFolder = \TYPO3\CMS\Core\Core\Environment::getProjectPath() . DIRECTORY_SEPARATOR. $this->input->getArgument('sourceFolder');
         $zipFile = $this->getFileInFolder($sourceFolder, 'zip');
         if ($zipFile !== false) {
             $this->output->writeln('Found zipfile: ' . $zipFile);
 
             $filename = basename($zipFile, '.zip');
-            $targetFolder = $sourceFolder . $filename . DIRECTORY_SEPARATOR;
+            $targetFolder = \TYPO3\CMS\Core\Core\Environment::getProjectPath() . DIRECTORY_SEPARATOR. $this->input->getArgument('targetFolder') . $filename . DIRECTORY_SEPARATOR;
 
             $this->prepareTargetFolder($targetFolder);
             $this->output->writeln('Extracting ' . $zipFile . ' to ' . $targetFolder);
@@ -90,10 +99,19 @@ class ImportCommand  extends Command
                     if ($this->input->getArgument('truncateTables')) {
                         $this->truncateOpenImmoTables();
                     }
+
+                    // import new data
                     $this->importFromXmlFile($xmlFilename);
+
+                    // adjust the paths of the files
+                    if ($this->input->getArgument('adjustAssetPaths')) {
+                        $this->adjustAssetPaths($targetFolder);
+                    }
                 }
-                $this->output->writeln('Deleting ' . $zipFile);
-                unlink($zipFile);
+                if ($this->input->getArgument('removeZipfile')) {
+                    $this->output->writeln('Deleting ' . $zipFile);
+                    unlink($zipFile);
+                }
             } else {
                 throw new \Exception($zipFile . ' could not be opened!');
             }
@@ -120,6 +138,30 @@ class ImportCommand  extends Command
                 $connectionPool->getConnectionForTable($tablename)->truncate($tablename);
             }
         }
+    }
+
+    /**
+     * Adds the path prefix to the filename.
+     *
+     * @param string $targetFolder
+     *
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    protected function adjustAssetPaths(string $targetFolder)
+    {
+        $targetFolder = str_replace(Environment::getPublicPath(), '', $targetFolder);
+
+        $this->output->writeln('Adjusting asset path to ' . $targetFolder);
+        /* @var $pool ConnectionPool */
+        $pool = $this->getObjectManager()->get(ConnectionPool::class);
+
+        $qb = $pool->getQueryBuilderForTable(self::ASSET_TABLE_NAME);
+        $qb
+            ->update(self::ASSET_TABLE_NAME)
+            // TODO: normally, we would just $qb->expr()->concat() here, but that doesn't work
+            ->set(self::ASSET_FIELD_NAME, $targetFolder)
+            ->execute()
+        ;
     }
 
     /**
@@ -218,9 +260,6 @@ class ImportCommand  extends Command
         if ( ! is_dir($targetFolder)) {
             $this->output->writeln('Creating target folder: ' . $targetFolder);
             GeneralUtility::mkdir($targetFolder);
-        } else {
-            $this->output->writeln('Cleaning target folder: ' . $targetFolder);
-            GeneralUtility::flushDirectory($targetFolder, true);
         }
     }
